+++ src/main/java/com/sprint/mission/discodeit/event/listener/kafka/NotificationRequiredTopicListener.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.event.listener.kafaka;
+package com.sprint.mission.discodeit.event.listener.kafka;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
